# perf(graphql-model-transformer): minimal provider framework and inline policies