… dependency_licenses.txt
… packages/amplify-category-api/src/__tests__/graphql-transformer/override/__snapshots__/model-transformer-override.test.ts.snap
… packages/amplify-graphql-model-transformer/API.md
… packages/amplify-graphql-model-transformer/package.json
… packages/amplify-graphql-model-transformer/src/__tests__/__snapshots__/amplify-dynamodb-table-generator.test.ts.snap
… packages/amplify-graphql-model-transformer/src/__tests__/amplify-dynamodb-table-generator.test.ts
… packages/amplify-graphql-model-transformer/src/__tests__/provider.test.ts
… packages/amplify-graphql-model-transformer/src/__tests__/util.test.ts
… packages/amplify-graphql-model-transformer/src/resources/amplify-dynamodb-table/amplify-table-manager-lambda/outbound.ts
… packages/amplify-graphql-model-transformer/src/resources/amplify-dynamodb-table/amplify-table-manager-lambda/util.ts
… packages/amplify-graphql-model-transformer/src/resources/amplify-dynamodb-table/provider.ts
… packages/amplify-graphql-model-transformer/src/resources/amplify-dynamodb-table/waiter-state-machine.ts
… packages/amplify-graphql-model-transformer/src/resources/dynamo-model-resource-generator.ts
… yarn.lock